requirements.txt
psutil
pandas
matplotlib
pytest
jinja2
pyyaml
blueutil